% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection.R
\docType{data}
\name{AeronetOC2019}
\alias{AeronetOC2019}
\title{Collection of coastal AeronetOC water spectra till 2019}
\format{
A dataframe with 10667 rows by 14 cols
}
\usage{
AeronetOC2019
}
\description{
A dataframe contains AeronetOC data
}
\details{
AeronetOC is collected from Level 2 products provided by Aeronet Ocean Color stations located
  in great inland lakes (i.e., Lake_Erie, Palgrunden) and coastal waters (i.e., COVE_SEAPRISM,
  Galata_Platform, Gloria, Gustav_Dalen_Tower, Helsinki_Lighthouse, LISCO, MVCO, and Zeebrugge-MOW1)
   with bands at 410, 440, 490, 530, 550, 667, 869 nm.
}
\note{
Colnames are "SampleID, X410nm, X440nm, X490nm, X530nm, 
  X550nm, X667nm, X869nm, X1020nm, Pressure, Wind_Speed, Chlorophyll.a, Sea_Surface_Reflectance, Ozone"
}
\references{
All these data are available at https://aeronet.gsfc.nasa.gov.
}
\concept{Datesets}
\keyword{datasets}
