% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{apply_FCM_m}
\alias{apply_FCM_m}
\title{Apply the improved Fuzzy Cluster Method (FCMm) to new data based on the defined centroids}
\usage{
apply_FCM_m(
  Rrs,
  wavelength = NULL,
  Rrs_clusters = NULL,
  m_used = 1.36,
  do.stand = FALSE,
  stand = NULL,
  default.cluster = "B2021_norm",
  quality_check = FALSE,
  option.plot = FALSE,
  color_palette = NULL
)
}
\arguments{
\item{Rrs}{Data.frame, the input Rrs of FCM.}

\item{wavelength}{Numeric vector, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{Rrs_clusters}{Data.frame, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{m_used}{Number, Used fuzzifier value}

\item{do.stand}{Logical, whether to normalized the Rrs data (both for Input and Centroids).
Default as \code{FALSE} means do not.}

\item{stand}{Deprecated; same to \code{do.stand}.}

\item{default.cluster}{Logical, whether to use the default clusters.
Default use the data from \code{Bi_clusters.rda}}

\item{quality_check}{Logical, quality chech option (default as \code{FALSE})}

\item{option.plot}{Logical, whether to plot the result. Default as \code{FALSE}}

\item{color_palette}{The palette of cluster color. Default as \code{RdYlBu(res$K)} if \code{NULL}.
It could be \code{RdYlBu(k)}, \code{Spectral(k)}, \code{HUE(k)} or other color values
with same length of cluster number. \code{k} means the cluster number.}
}
\value{
A \code{list} including several results of function \code{apply_FCM_m()}
  \itemize{
    \item \strong{x}  The raw input Rrs dataframe with unit sr^-1
    \item \strong{x.stand}  The standardized Rrs dataframe, if \code{stand=F}
    \item \strong{d}  Distance to each cluster
    \item \strong{u}  Membership values
    \item \strong{Area} Spectral intergration of each sample
    \item \strong{cluster}  Defined by the maximum of membership
    \item \strong{quality}  The quality of the cluster results.
    \item \strong{m.used}  The used value of fuzzifier(m)
    \item \strong{K}  Cluster number
    \item \strong{p.group}  A ggplot list for plotting the cluster result
    \item \strong{p.group.facet} Recommened! \code{p.group} with facet to see each cluster results
      more clearly
    \item \strong{dt.melt} Dataframe used for ggplot including \code{dt_plot_x}, \code{dt_plot_v}, and \code{cp}.
  }
}
\description{
Application of the improved Fuzzy Cluster Method (FCMm) for new Rrs data based on 
  default cluster settings or user-defined clusters (trained by FCM.new).
}
\examples{
library(FCMm)
Rrs <- WaterSpec35[,3:17]
# you have to make sure the training set and cluster are in the same manner
#   which means if cluster is on the normalized scale the the input should
#   be normalized before running it.
nRrs <- Rrs / trapz2(Rrs)
apply_FCM_m(nRrs, default.cluster = "B2021_norm", option.plot=TRUE)$p.group.facet
apply_FCM_m(nRrs, default.cluster = "BPHD_norm", option.plot=TRUE)$p.group.facet
apply_FCM_m(nRrs, default.cluster = "B2019_norm", option.plot=TRUE)$p.group.facet

}
\references{
Bi, S., Li, Y., Liu, G., Song, K., Xu, J., Dong, X., Cai, X., Mu, M., Miao, S., & Lyu, H. (2021). 
  Assessment of Algorithms for Estimating Chlorophyll-a Concentration in Inland Waters: 
  A Round-Robin Scoring Method Based on the Optically Fuzzy Clustering. IEEE Transactions on 
  Geoscience and Remote Sensing.

Bi, S., Li, Y., Xu, J., Liu, G., Song, K., Mu, M., Lyu, H., Miao, S., & Xu, J. (2019). 
  Optical classification of inland waters based on an improved Fuzzy C-Means method. 
  Optics Express, 27(24), 34838–34856.

Jackson, T., Sathyendranath, S., & Mélin, F. (2017). An improved optical classification 
  scheme for the Ocean Colour Essential Climate Variable and its applications. Remote 
  Sensing of Environment, 203, 152–161.

Moore, T. S., Dowell, M. D., Bradt, S., & Verdu, A. R. (2014). An optical water type 
  framework for selecting and blending retrievals from bio-optical algorithms in
  lakes and coastal waters. Remote Sensing of Environment, 143, 97–111.
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_to_image}()},
\code{\link{cal_memb}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
