% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{Assessment_via_cluster}
\alias{Assessment_via_cluster}
\title{Assessment each algorithm for every cluster}
\usage{
Assessment_via_cluster(
  pred,
  meas,
  memb,
  metrics = c("MAE", "MAPE"),
  log10 = FALSE,
  total = TRUE,
  hard.mode = TRUE,
  na.process = FALSE
)
}
\arguments{
\item{pred}{prediciton of Chla}

\item{meas}{in-situ measurement of Chla}

\item{memb}{membership value matrix}

\item{metrics}{metrics need to be calculated}

\item{log10}{Should pred and meas be log10-transformed? (Default as FALSE)}

\item{total}{logical}

\item{hard.mode}{If \code{FALSE}, the membership values are used to calculate validation metrics}

\item{na.process}{na.process and choose to statistic na value percent}
}
\value{
List
}
\description{
Assessment each algorithm for every cluster
}
\concept{Algorithm assessment}
