% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_blending.R
\name{Blend_FCMm}
\alias{Blend_FCMm}
\title{Algorithm blending framework by Bi theses}
\usage{
Blend_FCMm(Rrs, wv_range = 3, reparam = TRUE, ...)
}
\arguments{
\item{Rrs}{Rrs data.frame input for matching centroids and calculating Chla}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}

\item{reparam}{Whether to apply reparameterization coefficients}

\item{...}{parameters of \link{apply_FCM_m}}
}
\value{
A list includes \code{Chla_blend}, \code{Rrs}, \code{res_FCM} (the return of \link{apply_FCM_m}),
  and estimates from optimal algorithms.
}
\description{
Algorithm blending framework by Bi theses
}
\examples{
library(FCMm)
data(WaterSpec35)
# res = Blend_FCMm(WaterSpec35[, -c(1,2)])
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Bi21}()},
\code{\link{Blend_Jac17}()},
\code{\link{Blend_Moo14}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
