% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Blend_Jac17}
\alias{Blend_Jac17}
\title{Algorithm blending framework by Jackson et al. (2017)}
\usage{
Blend_Jac17(Rrs, wv_range = 5, ...)
}
\arguments{
\item{Rrs}{Rrs data.frame input for matching centroids and calculating Chla}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}

\item{...}{parameters of \link{apply_FCM_m}}
}
\value{
A list includes \code{Chla_blend}, \code{Rrs}, \code{res_FCM} (the return of \link{apply_FCM_m}),
  and estimates from optimal algorithms.
}
\description{
Algorithm blending framework by Jackson et al. (2017)
}
\examples{
library(FCMm)
data(WaterSpec35)
res = Blend_Jac17(WaterSpec35)
}
\references{
Jackson T, Sathyendranath S, Mélin F. An improved optical classification scheme for the Ocean 
  Colour Essential Climate Variable and its applications[J]. Remote Sensing of Environment, 2017, 
  203: 152-161.
  
Moore T S, Dowell M D, Bradt S, et al. An optical water type framework for 
  selecting and blending retrievals from bio-optical algorithms in lakes and coastal waters[J]. 
  Remote sensing of environment, 2014, 143: 97-111.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Moo14}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
