% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_blending.R
\name{Blend_Moo14}
\alias{Blend_Moo14}
\title{Algorithm blending framework by Moore et al. (2014)}
\usage{
Blend_Moo14(Rrs, wv_range = 3, m_used = 1.5, ...)
}
\arguments{
\item{Rrs}{Rrs data.frame input for matching centroids and calculating Chla}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}

\item{...}{parameters of \link{apply_FCM_m}}
}
\value{
A list includes \code{Chla_blend}, \code{Rrs}, \code{res_FCM} (the return of \link{apply_FCM_m}),
  and estimates from optimal algorithms.
}
\description{
Algorithm blending framework by Moore et al. (2014)
}
\examples{
library(FCMm)
data(WaterSpec35)
res = Blend_Moo14(WaterSpec35[, -c(1, 2)])
}
\references{
Moore T S, Dowell M D, Bradt S, et al. An optical water type framework for 
  selecting and blending retrievals from bio-optical algorithms in lakes and coastal waters[J]. 
  Remote sensing of environment, 2014, 143: 97-111.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Bi21}()},
\code{\link{Blend_FCMm}()},
\code{\link{Blend_Jac17}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
