% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{BR_Gil10}
\alias{BR_Gil10}
\title{BR_Gil10}
\usage{
BR_Gil10(Rrs665, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}
}
\description{
BR_Gil10
}
\references{
Gilerson A A, Gitelson A A, Zhou J, et al. Algorithms for remote estimation of 
  chlorophyll-a in coastal and inland waters using red and near infrared bands[J]. 
  Optics Express, 2010, 18(23): 24109-24125.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{CI_Hu12}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
