% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{C6}
\alias{C6}
\title{C6 (Being deprecated)}
\usage{
C6(Rrs665, Rrs754)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs754}{Rrs754}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
C6 (Being deprecated)
}
\examples{
data(WaterSpec35)
res = Bloom(WaterSpec35$`665`, WaterSpec35$`753.75`)
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
