% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cal.metrics}
\alias{cal.metrics}
\title{Assess the performance of algorithms}
\usage{
cal.metrics(x,y,name="all")
}
\arguments{
\item{x}{True value}

\item{y}{Estimated value}

\item{name}{The name of metrics}
}
\description{
Assess the performance of algorithms
}
\seealso{
Other Utils: 
\code{\link{.level_to_variable}()},
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()}
}
\concept{Utils}
