% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cal.metrics.vector}
\alias{cal.metrics.vector}
\title{Assess the performance of algorithms as vector}
\usage{
cal.metrics.vector(x,y,name="all",log10=FALSE)
}
\arguments{
\item{x}{True value OR Actual value}

\item{y}{Estimated value OR Predict value}

\item{name}{The name of metrics}

\item{log10}{Logical. Whether the input x and y should be log10-transformed}
}
\description{
Assess the performance of algorithms as vector
}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics}()},
\code{\link{level_to_variable}()},
\code{\link{lmodel2_}()},
\code{\link{trim_sd}()}
}
\concept{Utils}
