% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{cal_memb}
\alias{cal_memb}
\title{calculate membership values based on x, v, and m}
\usage{
cal_memb(x, v, m = 2)
}
\arguments{
\item{x}{Input matrix}

\item{v}{Input centroids}

\item{m}{Fuzzifier parameter. Default is 2.}
}
\value{
A list including x, v, m, u (membership), d (distance)
}
\description{
calculate membership values based on x, v, and m
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
