% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Chla_FBA}
\alias{Chla_FBA}
\alias{FBA_Le13}
\alias{FBA_Yang10}
\title{Four-bands Chla algorithm by Yang et al. (2010) and Le et al. (2013)}
\usage{
FBA_Le13(Rrs665, Rrs681, Rrs709)

FBA_Yang10(Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs681}{Rrs681}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
Four-bands Chla algorithm by Yang et al. (2010) and Le et al. (2013)
}
\examples{
data(WaterSpec35)
res_L = FBA_Le13(WaterSpec35$`665`, WaterSpec35$`681.25`, WaterSpec35$`708.75`)
res_Y = FBA_Yang10(WaterSpec35$`665`, WaterSpec35$`708.75`, WaterSpec35$`753.75`)
}
\references{
\itemize{
  \item Le C, Hu C, Cannizzaro J, et al. Evaluation of chlorophyll-a remote
  sensing algorithms for an optically complex estuary[J]. Remote Sensing of Environment, 
  2013, 129: 75-89.
  \item Yang W, Matsushita B, Chen J, et al. An enhanced three-band index for estimating 
  chlorophyll-a in turbid case-II waters: case studies of Lake Kasumigaura, Japan, and Lake 
  Dianchi, China[J]. IEEE Geoscience and Remote Sensing Letters, 2010, 7(4): 655-659.
}
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
