% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Chla_Gil10_Git11}
\alias{Chla_Gil10_Git11}
\alias{TBA_Gil10}
\alias{BR_Gil10}
\alias{TBA_Git11}
\alias{BR_Git11}
\title{Band-Ratio and Three-bands Chla algorithm by Gilerson et al. (2010) and Gitelson et al. (2011)}
\usage{
TBA_Gil10(Rrs665, Rrs709, Rrs754)

BR_Gil10(Rrs665, Rrs709)

TBA_Git11(Rrs665, Rrs709, Rrs754)

BR_Git11(Rrs665, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
Band-Ratio and Three-bands Chla algorithm by Gilerson et al. (2010) and Gitelson et al. (2011)
}
\examples{
data(WaterSpec35)
res_TBA_Gil10 = TBA_Gil10(WaterSpec35$`665`, WaterSpec35$`708.75`, WaterSpec35$`753.75`)
res_BR_Gil10  = BR_Gil10(WaterSpec35$`665`, WaterSpec35$`708.75`)
res_TBA_Git11 = TBA_Git11(WaterSpec35$`665`, WaterSpec35$`708.75`, WaterSpec35$`753.75`)
res_BR_Git11 = BR_Git11(WaterSpec35$`665`, WaterSpec35$`708.75`)
}
\references{
\itemize{
  \item Gilerson A A, Gitelson A A, Zhou J, et al. Algorithms for remote estimation of 
    chlorophyll-a in coastal and inland waters using red and near infrared bands[J].
    Optics Express, 2010, 18(23): 24109-24125.
  \item Gitelson A A, Gurlin D, Moses W J, et al. Remote Estimation of Chlorophyll-a 
    Concentration[J]. Advances in environmental remote sensing: sensors, algorithms, 
    and applications, 2011: 439.
}
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
