% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Chla_OC_OLCI}
\alias{Chla_OC_OLCI}
\alias{OC4_OLCI}
\alias{OC5_OLCI}
\alias{OC6_OLCI}
\title{NASA standard ocean color algorithm (version 4, 5, and 6) for the OLCI bands}
\usage{
OC4_OLCI(Rrs443, Rrs490, Rrs510, Rrs560)

OC5_OLCI(Rrs412, Rrs443, Rrs490, Rrs510, Rrs560)

OC6_OLCI(Rrs412, Rrs443, Rrs490, Rrs510, Rrs560, Rrs665)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs490}{Rrs490}

\item{Rrs510}{Rrs510}

\item{Rrs560}{Rrs560}

\item{Rrs412}{Rrs412}

\item{Rrs665}{Rrs665}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
NASA standard ocean color algorithm (version 4, 5, and 6) for the OLCI bands
}
\examples{
data(WaterSpec35)
res_OC4 = OC4_OLCI(WaterSpec35$`442.5`, WaterSpec35$`490`, 
WaterSpec35$`510`, WaterSpec35$`560`)
res_OC5 = OC5_OLCI(WaterSpec35$`412.5`, WaterSpec35$`442.5`, 
WaterSpec35$`490`, WaterSpec35$`510`, WaterSpec35$`560`)
res_OC6 = OC6_OLCI(WaterSpec35$`412.5`, WaterSpec35$`442.5`, 
WaterSpec35$`490`, WaterSpec35$`510`, WaterSpec35$`560`, WaterSpec35$`665`)

}
\references{
O'Reilly J E, Werdell P J. Chlorophyll algorithms for ocean color
  sensors-OC4, OC5 & OC6[J]. Remote sensing of environment, 2019, 229: 32-47.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
