% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Color_plattes}
\alias{Color_plattes}
\alias{Spectral}
\alias{RdYlBu}
\alias{HUE}
\title{Color plattes including Spectral and RdYlBu}
\usage{
Spectral(n)

RdYlBu(n)

HUE(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
Color codes
}
\description{
Color plattes including Spectral and RdYlBu
}
\examples{
Spectral(7)
RdYlBu(7)
}
