% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection.R
\docType{data}
\name{dt_water}
\alias{dt_water}
\title{Absorption and scattering coefficient of pure water}
\format{A dataframe with 561 rows by 4 cols}
\usage{
dt_water
}
\description{
Absorption and scattering coefficient of pure water with unit m^-1
}
\details{
The wavelength is from 340 nm to 900 nm.
}
\note{
Will add later.
}
\references{
Will add later.
}
\seealso{
Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{OLCI_TH}},
\code{\link{Rrs_clusters.default}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{WaterSpec35}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
