% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{FBA_Le13}
\alias{FBA_Le13}
\title{FBA_Le13}
\usage{
FBA_Le13(Rrs665, Rrs681, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs681}{Rrs681}

\item{Rrs709}{Rrs709}
}
\description{
FBA_Le13
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4E}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
