% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{FBA_Yang10}
\alias{FBA_Yang10}
\title{FBA_Yang10}
\usage{
FBA_Yang10(Rrs665, Rrs708, Rrs753)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs708}{Rrs708}

\item{Rrs753}{Rrs753}
}
\description{
FBA_Yang10
}
