% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM.new.R
\name{FCM.new}
\alias{FCM.new}
\title{FCM.new}
\usage{
FCM.new(FDlist, K, plot.jitter=TRUE, fast.mode=FALSE, stand=FALSE)
}
\arguments{
\item{FDlist}{A \code{list} from function \code{\link{FuzzifierDetermination}}}

\item{K}{Number, cluster number}

\item{plot.jitter}{Logical, choose to plot jitter plot using \code{ggplot2} functions}

\item{fast.mode}{Logical, \code{FALSE} (default)}

\item{stand}{Logical, choose to do normalization of input data}
}
\value{
A \code{list} of FCM:
  \itemize{
    \item \strong{FD} The return list by function \code{\link{FuzzifierDetermination}}
    \item \strong{res.FCM} The optimized FCM result generated by functions in package \code{ppclust}
    \item \strong{K} Cluster number
    \item \strong{plot.jitter} A logical value for the option of doing jitter plot by package \code{ggplot2}
    \item \strong{fast.mode} A logical value for choosing whether to use fast mode
  }
}
\description{
An improved version of FCM for water spectra data sets.
}
\examples{
\dontrun{
library(FCMm)
library(tidyverse)
data("Nechad2015")
w <- Nechad2015 \%>\% names \%>\%
    str_extract(.,pattern="\\\\d") \%>\%
    is.na \%>\% {!.}
wv <- w \%>\% names(Nechad2015)[.] \%>\%
    gsub('X','',.) \%>\% as.numeric
x <- w \%>\% Nechad2015[,.]
names(x) <- wv
rm(w)
FD <- FuzzifierDetermination(x, wv, stand=F)
nb <- 4
set.seed(54321)
result <- FCM.new(FD, nb)
print(result$p.jitter)
}

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Dembele D. Multi-objective optimization for clustering 3-way gene
    expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3):
    211-225.
}
}
