% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{FuzzifierDetermination}
\alias{FuzzifierDetermination}
\title{Determine the optimized fuzzifier for Fuzzy Cluster Method (FCM) running}
\usage{
FuzzifierDetermination(x, wv, max.m=10, do.stand=TRUE, stand=NULL, dmetric="sqeuclidean")
}
\arguments{
\item{x}{Data.frame. the input Rrs data}

\item{wv}{Wavelength of X}

\item{max.m}{Set max.m as for determination of m.mub. Default as 10}

\item{do.stand}{Whether run standarization for the input data set. Default as \code{TRUE}.
Note that the do.stand only be used for calculating the fuzzifier as both raw and standarzied 
spectra will be restored in the return value. This is benifit to the spectra plotting.}

\item{stand}{Deprecated; Now \code{stand = !do.stand}}

\item{dmetric}{Distance method. Default as 'sqeuclidean'}
}
\value{
\code{FD} list contains several result by \code{FuzzifierDetermination}:
  \itemize{
    \item \strong{x} The raw input Rrs dataframe with unit sr^-1
    \item \strong{x.stand} The standardized Rrs dataframe, if \code{do.stand=TRUE}
    \item \strong{wv} Wavelength with unit nm
    \item \strong{max.m} The maximum fuzzifier of FCM as a restriction
    \item \strong{do.stand} A logic value for whether we standardized the input data
    \item \strong{dmetric} A string value for choosing which distance metric to be used
    \item \strong{Area} A numeric vector for trapezoidal integral values
    \item \strong{m.ub} The upper boundary of fuzzifier(m) value
    \item \strong{m.used} The desired value of fuzzifier(m) value
  }
}
\description{
To determine the optimized fuzzifier value for Fuzzy Cluster Method (FCM)running.
}
\examples{

library(FCMm) 
library(magrittr)
data("Nechad2015")
x <- Nechad2015[,3:11]
wv <- gsub("X","",names(x)) \%>\% as.numeric
set.seed(1234)
w <- sample(1:nrow(x), 100) 
x <- x[w, ]
names(x) <- wv
set.seed(1234)
FD <- FuzzifierDetermination(x, wv, stand=FALSE)

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Dembele D. Multi-objective optimization for clustering 3-way gene
    expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3):
    211-225.
}
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
