% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{FuzzifierDetermination2}
\alias{FuzzifierDetermination2}
\title{The version 2 of FuzzifierDetermination}
\usage{
FuzzifierDetermination2(x, iter.max = 300, m_ini_method = 2)
}
\arguments{
\item{x}{data.frame x.}

\item{iter.max}{the max of iteration number. defualt as 300.}

\item{m_ini_method}{the initialization m value. 2 (default) or "Schwammle2010".}
}
\value{
m_ub and m_used
}
\description{
The version 2 of FuzzifierDetermination
}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Dembele D. Multi-objective optimization for clustering 3-way gene
    expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3):
    211-225.
  \item Schwammle V, Jensen O N. A simple and fast method to determine the parameters for 
    fuzzy c–means cluster analysis[J]. Bioinformatics, 2010, 26(22): 2841-2848.
}
}
