% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSM_estimation.R
\name{GAA_SPM}
\alias{GAA_SPM}
\title{Globally applicable algorithm to seamlessly retrieve the 
  concentration of suspended particulate matter from remote sensing reflectance.}
\usage{
GAA_SPM(Rrs490, Rrs560, Rrs674, Rrs754, Rrs865)
}
\arguments{
\item{Rrs490}{Rrs490}

\item{Rrs560}{Rrs560}

\item{Rrs674}{Rrs674}

\item{Rrs754}{Rrs754}

\item{Rrs865}{Rrs865}
}
\value{
The result returned by \code{GAA_SPM} is a list including:
\itemize{
  \item \strong{param} Coefficients of C0, C1, C2, C3, a1, a2.
  \item \strong{W} The weighting factors (data.frame).
  \item \strong{GI} Generalized index for SPM.
  \item \strong{SPM} The predicted suspended particulate matter.
}
}
\description{
Globally applicable algorithm to seamlessly retrieve the 
  concentration of suspended particulate matter from remote sensing reflectance.
}
\note{
The input wavelength are at OLCI bandset but this function will
  transform these wavelengthes to the GAA required bands.
}
\references{
Yu X, Lee Z, Shen F, et al. An empirical algorithm to seamlessly retrieve the 
  concentration of suspended particulate matter from water color across ocean to 
  turbid river mouths[J]. Remote Sensing of Environment, 2019, 235: 111491.
}
