% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_application 2.R, R/Image_application.R
\name{generate_param}
\alias{generate_param}
\title{generate_param}
\usage{
generate_param(wl)

generate_param(wl)
}
\arguments{
\item{wl}{wavelength of subsetting}
}
\description{
generate_param

generate_param
}
