% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{level_to_variable}
\alias{level_to_variable}
\title{Convert dataframe with factor to character}
\usage{
level_to_variable(dt, warn = FALSE)
}
\arguments{
\item{dt}{dataframe}

\item{warn}{warning option}
}
\value{
The dataframe with the leveled columns converted to the character columns.
}
\description{
Convert dataframe with factor to character
}
\examples{
x = data.frame(x=seq(1,10), y=as.character(rep(c(1,2),5)))
x_ = level_to_variable(x)

}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{trim_sd}()}
}
\concept{Utils}
