% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lmodel2_}
\alias{lmodel2_}
\title{Run SMA linear regression}
\usage{
lmodel2_(formula, data = NULL, range.y = NULL, range.x = NULL, nperm = 0)
}
\arguments{
\item{formula}{See package \code{lmodel2} for more details}

\item{data}{See package \code{lmodel2} for more details}

\item{range.y}{See package \code{lmodel2} for more details}

\item{range.x}{See package \code{lmodel2} for more details}

\item{nperm}{See package \code{lmodel2} for more details}
}
\value{
A lmodel2 function list
}
\description{
Run SMA linear regression
}
\note{
This function was modified by the package \code{lmodel2} (Version 1.7-3) since 
  the message as output is unnecessary in most of scene. I just remove this messages 
  in by comment.
}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{level_to_variable}()},
\code{\link{trim_sd}()}
}
\concept{Utils}
