% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection.R
\docType{data}
\name{Nechad2015}
\alias{Nechad2015}
\title{Collection of in_situ data by Nechad et al. (2015)}
\format{
A dataframe with 336 rows by 18 cols
}
\usage{
Nechad2015
}
\description{
A dataframe contains Nechad2015 data
}
\details{
None
}
\note{
Colnames are "DataProvider, SAMPLE.ID, X412.5, X442.5, X490, X510, X560, X620,
  X665, X681.25, X708.75, X.Chl_a..ug.L., X.TSM..mg.l., CC_SITE, DATE.dd.mm.yyyy.,
 time, latitude, longitude"
}
\references{
B. Nechad, K. Ruddick, T. Schroeder, K. Oubelkheir, D. Blondeau-Patissier, N. Cherukuru, 
  V. Brando, A. Dekker, L. Clementson, and A. C. Banks, "CoastColour Round Robin data sets: a database to
  evaluate the performance of algorithms for the retrieval of water quality parameters in coastal waters,"
  Earth system science data 7, 319-348 (2015). (https://doi.pangaea.de/10.1594/PANGAEA.841950)
}
\seealso{
Other Datasets: 
\code{\link{OLCI_TH}},
\code{\link{Rrs_clusters.default}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{WaterSpec35}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
