% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{OC4E}
\alias{OC4E}
\title{OC4E}
\usage{
OC4E(Rrs443, Rrs490, Rrs510, Rrs555)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs490}{Rrs490}

\item{Rrs510}{Rrs510}

\item{Rrs555}{Rrs555}
}
\description{
OC4E
}
