% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{OCI_Hu12}
\alias{OCI_Hu12}
\title{OCI_Hu12}
\usage{
OCI_Hu12(Rrs443, Rrs490, Rrs510, Rrs560, Rrs665)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs490}{Rrs490}

\item{Rrs510}{Rrs510}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}
}
\description{
OCI_Hu12
}
\references{
Hu C, Lee Z, Franz B. Chlorophyll aalgorithms for oligotrophic oceans: 
  A novel approach based on three‐band reflectance difference[J]. Journal of Geophysical 
  Research: Oceans, 2012, 117(C1).
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
