% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection.R
\docType{data}
\name{OLCI_TH}
\alias{OLCI_TH}
\title{A test Ocean and Land Color Instrument (OLCI) image raster in Lake Taihu}
\format{
A RasterBrick with 791136 elements
}
\usage{
OLCI_TH
}
\description{
A dataframe contains OLCI_TH data
}
\details{
This dataset is a OLCI raster in Lake Taihu for testing the package `FCMm`.
}
\note{
The dataset should be only used in this package test. If any want 
  to use it in your own study, he or she should contact us via bishun1994 at foxmail or 
  liyunmei at njnu.edu.cn
  
  This OLCI image onboard Sentinel-3A (ESA) in Lake Taihu was scened on 17 August, 2018
}
\references{
* Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  
* Liu G, Li Y, Lyu H, et al. An improved land target-based atmospheric correction method 
  for Lake Taihu[J]. IEEE Journal of Selected Topics in Applied Earth Observations and
  Remote Sensing, 2015, 9(2): 793-803.
}
\seealso{
Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{SRF_LIST}},
\code{\link{Spectra.Centroids}},
\code{\link{Valente2019}},
\code{\link{WaterSpec35}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
