% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{plot_spec}
\alias{plot_spec}
\title{Plot the result of FCM_m}
\usage{
plot_spec(res, show.stand=FALSE, HABc=NULL)
}
\arguments{
\item{res}{The result of function `FCM.new`}

\item{show.stand}{Logical option to show the standardized spectra (default as False)}

\item{HABc}{Numeric value to specify which cluster owning larger magnitude}
}
\value{
A list of ggplot lists:
  \itemize{
    \item \strong{p.all.spec} Spectra of all data.
    \item \strong{p.cluster.spec} Spectra of clusters.
    \item \strong{p.group.spec.1} Spectra of groups in mode 1
    \item \strong{p.group.spec.2} Spectra of groups in mode 2
    \item \strong{res} Result list of \code{FCM.new}
    \item \strong{show.stand} Logical value whether use standardized y-axis
  }
}
\description{
Spectra plot for a FCM.new result
}
\examples{
\dontrun{
library(FCMm) 
library(ggplot2) 
library(magrittr)
data("Nechad2015")
x <- Nechad2015[,3:11]
wv <- gsub("X","",names(x)) \%>\% as.numeric
set.seed(1234) # Set this seed so that you can re-produce them
w <- sample(1:nrow(x), 100)
x <- x[w, ]
names(x) <- wv
nb = 4 # Obtained from the vignette "Cluster a new dataset by FCMm"
set.seed(1234)
FD <- FuzzifierDetermination(x, wv, stand=FALSE)
result <- FCM.new(FD, nb, fast.mode = TRUE)
p.spec <- plot_spec(result, show.stand=TRUE, HABc=NULL)
print(p.spec$p.cluster.spec)
}

}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
