% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{plot_spec_from_df}
\alias{plot_spec_from_df}
\title{plot_spec_from_df}
\usage{
plot_spec_from_df(df)
}
\arguments{
\item{df}{Data.frame for Rrs plotting}
}
\value{
A ggplot list
}
\description{
Spectra plot for a dataframe given numeric colnames
}
\note{
The colnames of input `df` should be numericalize (i.e. names(df) %>% as.numeric)
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()}
}
\concept{Fuzzy cluster functions}
