% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{plot_spec_group}
\alias{plot_spec_group}
\title{Plot spectra from dataframe by group}
\usage{
plot_spec_group(x, group, palette = RdYlBu, facet = TRUE, group_num = TRUE)
}
\arguments{
\item{x}{The input matrix with colnames could be converted to numeric values.}

\item{group}{The vector indicate the group-belongings for x}

\item{palette}{Color palette function. The default is \link{RdYlBu}}

\item{facet}{Whether to plot by \link{facet_wrap}. The default is \code{TRUE}.}

\item{group_num}{Whether to add sample numbers of each group in strip.text. Also support 
character as the input.}
}
\value{
A ggplot list.
}
\description{
This function will help you quick plot the spectra if you have a data.frame
  that could be used for \link{plot_spec_from_df} and have a \code{group} parameter which indicates
  the different groups for spectra in your \code{x}.
}
