% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAA_series.R
\name{QAA_v5}
\alias{QAA_v5}
\title{QAA_v5}
\usage{
QAA_v5(
  wv,
  Rrs,
  wv412 = NULL,
  wv443 = NULL,
  wv490 = NULL,
  wv560 = NULL,
  wv667 = NULL,
  verbose = F
)
}
\arguments{
\item{wv}{wv}

\item{Rrs}{Rrs}

\item{wv412}{Wavelength index of 412 nm}

\item{wv443}{Wavelength index of 443 nm}

\item{wv490}{Wavelength index of 490 nm}

\item{wv560}{Wavelength index of 560 nm}

\item{wv667}{Wavelength index of 667 nm}

\item{verbose}{verbos (default as FASLE)}
}
\description{
QAA_v5
}
\references{
Lee Z P, Carder K L, Arnone R A. Deriving inherent optical properties from water color: 
  a multiband quasi-analytical algorithm for optically deep waters[J]. Applied optics, 2002, 41(27): 
  5755-5772.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
