% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rrs_clusters.default.R
\docType{data}
\name{Rrs_clusters.default}
\alias{Rrs_clusters.default}
\title{Seven built-in clusters}
\format{dataframe with 7 rows by 15 cols}
\usage{
Rrs_clusters.default
}
\description{
A dataframe contains the water optical clusters produced by Bi et al. (2019)
}
\note{
You have to take a look at the description of this built-in cluster (water type) such as
  thier corresponding water color parameters in Bi et al. (2019).
}
\seealso{
Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{OLCI_TH}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{WaterSpec35}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
