% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{run_all_Chla_algorithms}
\alias{run_all_Chla_algorithms}
\alias{Chla_algorithms_name}
\title{Run all Chla algorithms}
\usage{
run_all_Chla_algorithms(Rrs, wv_range = 3)

Chla_algorithms_name()
}
\arguments{
\item{Rrs}{Dataframe that should with required colnames 443, 490, 510, 560, 620, 665, 681, 709, 754, 779}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}
}
\value{
A list including Chla results by supported algorithms if Rrs includes the required wavelength.
  Retrun \code{NA} values for algorithms of which required wavelength cannot be found.
}
\description{
Run all Chla algorithms
}
\details{
Run \link{Chla_algorithms_name} to see all Chla algorithms
}
\examples{
library(FCMm)
data(WaterSpec35)
out = run_all_Chla_algorithms(WaterSpec35[,-c(1,2)])

}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()}
}
\concept{Algorithms: Chla concentration}
