% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{Sampling_via_cluster}
\alias{Sampling_via_cluster}
\alias{Sampling_by_sort}
\title{Stratified sampling by clusters or types}
\usage{
Sampling_via_cluster(
  x,
  num = length(x),
  replace = FALSE,
  order.value = NULL,
  ...
)

Sampling_by_sort(
  x,
  num = length(x),
  log10 = FALSE,
  n_group = 3,
  replace = TRUE
)
}
\arguments{
\item{x}{A vector represents the cluster or type, only support numeric value now.}

\item{num}{The number of sampling. Default as \code{length(x)}}

\item{replace}{The way of sampling. See \code{help(sample)} for more details.}

\item{order.value}{The vector used for ordering. Default is \code{NULL}.}

\item{...}{Parameters of function \code{Sampling_by_sort}.}

\item{log10}{Whether to log10-trans the \code{order.value}. Default is \code{FALSE}.}

\item{n_group}{Number of group to be devided. Default is 3.}
}
\value{
Result of \code{Sampling_via_cluster} is the index of sampled 
  items from the input \code{x}.
}
\description{
Stratified sampling by clusters or types
}
\examples{

library(FCMm)
x = round(runif(100, 1, 10))
table(x)
w_sampled = Sampling_via_cluster(x, 20)
table(x[w_sampled])


x   <- stats::rnorm(100)
ind <- Sampling_by_sort(x)

}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Score_algorithms_interval}()},
\code{\link{Score_algorithms_sort}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
