% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{SCI_Shen10}
\alias{SCI_Shen10}
\title{SCI_Shen10}
\usage{
SCI_Shen10(Rrs560, Rrs620, Rrs665, Rrs681)
}
\arguments{
\item{Rrs560}{Rrs560}

\item{Rrs620}{Rrs620}

\item{Rrs665}{Rrs665}

\item{Rrs681}{Rrs681}
}
\description{
SCI_Shen10
}
