% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Score_algorithms_interval}
\alias{Score_algorithms_interval}
\title{Score a vector of error metrics for algorithms by the interval}
\usage{
Score_algorithms_interval(
  x,
  trim = FALSE,
  reward.punishment = TRUE,
  decreasing = TRUE,
  hundred.percent = FALSE
)
}
\arguments{
\item{x}{Input vector of error metrics (NA values are allowed)}

\item{trim}{whether to run a trim process to calculate mean and standard deviation of 
input vector x (Default as \code{FALSE})}

\item{reward.punishment}{Whether to conduct the reward and punishment mechanism in scoring
system (Default as \code{TRUE})}

\item{decreasing}{the order of the good metric to be evaluated. For instance, MAE should use
\code{decreasing = TRUE} (Default) since the algorithm performs better when MAE becomes smaller. 
However, when comes to \code{Rsquare} from linear regression (maximum is 1), it should be
\code{FALSE}}

\item{hundred.percent}{A variable constrain for metrics that the maximun of input \code{x} 
should not be greater than 100.}
}
\value{
Results of \code{Score_algorithms_interval()} are returned as a list including:
\item{p}{A ggplot list of the scoring result.}
\item{score}{The final score from by the interval score.}
\item{u}{Trimmed mean of input x with \code{NA} values removed.}
\item{bds}{Up and low boundaries for determining scores.}
\item{x}{The input x.}
}
\description{
Score a vector of error metrics for algorithms by the interval
}
\examples{
\dontrun{
set.seed(1234)
x = runif(10)
result = Score_algorithms_interval(x)
}
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_sort}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
