% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment 2.R,
%   R/Algorithms_assessment.R
\name{Scoring_system}
\alias{Scoring_system}
\title{The main function for algorithms scoring}
\usage{
Scoring_system(
  Inputs,
  method = "sort-based",
  param_sort = list(decreasing = TRUE),
  param_interval = list(trim = FALSE, reward.punishment = TRUE, decreasing = TRUE,
    hundred.percent = FALSE),
  remove.negative = FALSE
)

Scoring_system(
  Inputs,
  method = "sort-based",
  param_sort = list(decreasing = TRUE),
  param_interval = list(trim = FALSE, reward.punishment = TRUE, decreasing = TRUE,
    hundred.percent = FALSE),
  remove.negative = FALSE
)
}
\arguments{
\item{Inputs}{The list returned form function \code{Getting_Asses_results}}

\item{method}{The method selected to score algorithms: 'sort-based' (default) or 'interval-based'}

\item{param_sort}{The parameters of function \code{Score_algorithms_sort}}

\item{param_interval}{The parameters of function \code{Score_algorithms_interval}}

\item{remove.negative}{Option to replace the negative score as zero (Default as \code{FALSE})}
}
\value{
A list including Total_score, Accuracy, Precision, Effectiveness, Accuracy.list, 
  Precision.list, and Total_score.melt.

A list including Total_score, Accuracy, Precision, Effectiveness, Accuracy.list, 
  Precision.list, and Total_score.melt.
}
\description{
The main function for algorithms scoring

The main function for algorithms scoring
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_interval}()}

Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_interval}()}
}
\concept{Algorithm assessment}
