% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{show_sensor_names}
\alias{show_sensor_names}
\title{show_sensor_names}
\usage{
show_sensor_names()
}
\value{
A character string showing what sensors are supported in the package.
}
\description{
show_sensor_names
}
\examples{
\dontrun{show_sensor_names()}
}
