% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Smith18}
\alias{Smith18}
\title{Smith algorithm by Smith et al. (2018)}
\usage{
Smith18(Rrs443, Rrs490, Rrs510, Rrs560, Rrs665, Rrs709)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs490}{Rrs490}

\item{Rrs510}{Rrs510}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}
}
\value{
A list including the algorithm estimated CHL_G2B, CHL_OCI, 
  CHL_BLEND concentration. PHI index. a1 and a2 coefficients.
}
\description{
Smith algorithm by Smith et al. (2018)
}
\examples{
data(WaterSpec35)
res = Smith18(WaterSpec35$`442.5`, WaterSpec35$`490`, WaterSpec35$`510`, 
WaterSpec35$`560`, WaterSpec35$`665`, WaterSpec35$`708.75`)
}
\references{
Smith M E, Lain L R, Bernard S. An optimized chlorophyll a switching 
  algorithm for MERIS and OLCI in phytoplankton-dominated waters[J]. 
  Remote Sensing of Environment, 2018, 215: 217-227.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
