% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_centroids.R
\docType{data}
\name{Spectra.Centroids}
\alias{Spectra.Centroids}
\title{Several built-in spectral centroids}
\usage{
data("B2019_norm")
data("B2021_norm")
data("BPHD_norm")
data("Jac17_raw")
data("Moo14_raw")
}
\description{
A data frame contains the water optical clusters produced by Bi et al. (2019),
  Bi et al. (2021), Bi's PhD thesis paper in 2021, Jackson et al. (2017), and Moore et al. (2014)
}
\note{
\code{Bi2019_norm}, \code{Bi2021_norm}, \code{BPHD_norm} are normalized centroids 
  based on spectral integrals.

\code{Jac17_raw} is the Rrs centroids based on the raw scale (using Mahalanobis distance)

\code{Moo14_raw} is converted from the Rrs(0-) by using Rrs(0+) = 0.52 / (1/Rrs(0-) - 1.7)
}
\references{
Bi, S., Li, Y., Liu, G., Song, K., Xu, J., Dong, X., Cai, X., Mu, M., Miao, S., & Lyu, H. (2021). 
  Assessment of Algorithms for Estimating Chlorophyll-a Concentration in Inland Waters: 
  A Round-Robin Scoring Method Based on the Optically Fuzzy Clustering. IEEE Transactions on 
  Geoscience and Remote Sensing.

Bi, S., Li, Y., Xu, J., Liu, G., Song, K., Mu, M., Lyu, H., Miao, S., & Xu, J. (2019). 
  Optical classification of inland waters based on an improved Fuzzy C-Means method. 
  Optics Express, 27(24), 34838–34856.

Jackson, T., Sathyendranath, S., & Mélin, F. (2017). An improved optical classification 
  scheme for the Ocean Colour Essential Climate Variable and its applications. Remote 
  Sensing of Environment, 203, 152–161.

Moore, T. S., Dowell, M. D., Bradt, S., & Verdu, A. R. (2014). An optical water type 
  framework for selecting and blending retrievals from bio-optical algorithms in
  lakes and coastal waters. Remote Sensing of Environment, 143, 97–111.
}
\seealso{
Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{OLCI_TH}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{WaterSpec35}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
