% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{TBA_Gil10}
\alias{TBA_Gil10}
\title{TBA_Gil10}
\usage{
TBA_Gil10(Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\description{
TBA_Gil10
}
