% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{TBA_Git11}
\alias{TBA_Git11}
\title{Three-bands algorithm by Gitelson et al. (2011)}
\usage{
TBA_Git11(Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
Three-bands algorithm by Gitelson et al. (2011)
}
\references{
Gitelson A A, Gurlin D, Moses W J, et al. Remote Estimation of 
  Chlorophyll-a Concentration[J]. Advances in environmental remote sensing: sensors, 
  algorithms, and applications, 2011: 439.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
