% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{TC2_turbid}
\alias{TC2_turbid}
\title{TC2_turbid}
\usage{
TC2_turbid(Rrs443, Rrs560, Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
Chla concentration.
}
\description{
TC2_turbid
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2_clean}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
