% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{trim_sd}
\alias{trim_sd}
\title{Standard deviation by trimmed values}
\usage{
trim_sd(x, trim = 0.05, na.rm = TRUE)
}
\arguments{
\item{x}{input numeric vector}

\item{trim}{percent of length(x) to be trimmed}

\item{na.rm}{default as TRUE to remove NA values}
}
\value{
Trimmed standard value of the input data.
}
\description{
Standard deviation by trimmed values
}
\examples{
set.seed(1234)
x = runif(100)
sd(x)
trim_sd(x)

}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{level_to_variable}()}
}
\concept{Utils}
