% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection 2.R, R/Data_collection.R
\docType{data}
\name{WaterSpec35}
\alias{WaterSpec35}
\title{A test-purpose collection of inland water data}
\format{
A dataframe with 35 rows by 17 cols

A dataframe with 35 rows by 17 cols
}
\usage{
WaterSpec35

WaterSpec35
}
\description{
A dataframe contains WaterSpec35 data

A dataframe contains WaterSpec35 data
}
\details{
This dataset is a part of research of Bi et al. (2019) for
  testing the package `FCMm`. Colnames are "SampleID, Chla,
  400, 412.5, 442.5, 490, 510, 560, 620,
  665, 673.75, 681.25, 708.75, 753.75, 778.75, 865, 885"

This dataset is a part of research of Bi et al. (2019) for
  testing the package `FCMm`. Colnames are "SampleID, Chla,
  400, 412.5, 442.5, 490, 510, 560, 620,
  665, 673.75, 681.25, 708.75, 753.75, 778.75, 865, 885"
}
\note{
The dataset should be only used in this package test. If any want 
  to use it in your own study, he or she should contact us via bishun1994 at foxmail or 
  liyunmei at njnu.edu.cn

The dataset should be only used in this package test. If any want 
  to use it in your own study, he or she should contact us via bishun1994 at foxmail or 
  liyunmei at njnu.edu.cn
}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
\seealso{
Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{OLCI_TH}},
\code{\link{Rrs_clusters.default}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}

Other Datasets: 
\code{\link{Nechad2015}},
\code{\link{OLCI_TH}},
\code{\link{Rrs_clusters.default}},
\code{\link{SRF_LIST}},
\code{\link{Valente2019}},
\code{\link{dt_water}},
\code{\link{wavelength.default}}
}
\concept{Datasets}
\keyword{datasets}
