% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.default.R
\docType{data}
\name{wavelength.default}
\alias{wavelength.default}
\title{wavelength.default}
\format{a vector with 15 elements}
\usage{
wavelength.default
}
\description{
A vector contains the wavelength of water optical clusters produced by Bi et al. (2019)
}
\keyword{datasets}
