% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{apply_FCM_m}
\alias{apply_FCM_m}
\title{Apply FCM_m (improved Fuzzy Cluster Method) to new input Rrs}
\usage{
apply_FCM_m(
  Rrs,
  wavelength = NULL,
  Rrs_clusters = NULL,
  m_used = 1.36,
  stand = FALSE,
  default.cluster = TRUE,
  quality_check = FALSE,
  option.plot = FALSE,
  color_palette = NULL
)
}
\arguments{
\item{Rrs}{Data.frame, the input Rrs of FCM.}

\item{wavelength}{Numeric vector, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{Rrs_clusters}{Data.frame, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{m_used}{Number, Used fuzzifier value}

\item{stand}{Logical, whether to normalized the Rrs data (both for Input and Centroids).
Default as \code{FALSE} means do not.}

\item{default.cluster}{Logical, whether to use the default clusters.
Default use the data from \code{Bi_clusters.rda}}

\item{quality_check}{Logical, quality chech option (default as \code{FALSE})}

\item{option.plot}{Logical, whether to plot the result. Default as \code{FALSE}}

\item{color_palette}{The palette of cluster color. Default as \code{RdYlBu(res$K)} if \code{NULL}.
It could be \code{RdYlBu(k)}, \code{Spectral(k)}, \code{HUE(k)} or other color values
with same length of cluster number. \code{k} means the cluster number.}
}
\value{
A \code{list} including several results of function \code{apply_FCM_m()}
  \itemize{
    \item \strong{x}  The raw input Rrs dataframe with unit sr^-1
    \item \strong{x.stand}  The standardized Rrs dataframe, if \code{stand=F}
    \item \strong{d}  Distance to each cluster
    \item \strong{u}  Membership values
    \item \strong{Area} Spectral intergration of each sample
    \item \strong{cluster}  Defined by the maximum of membership
    \item \strong{quality}  The quality of the cluster results.
    \item \strong{m.used}  The used value of fuzzifier(m)
    \item \strong{K}  Cluster number
    \item \strong{p.group}  A ggplot list for plotting the cluster result
    \item \strong{p.group.facet} Recommened! \code{p.group} with facet to see each cluster results
      more clearly
    \item \strong{dt.melt} Dataframe used for ggplot including \code{dt_plot_x}, \code{dt_plot_v}, and \code{cp}.
  }
}
\description{
Application of FCM_m (improved Fuzzy Cluster Method) method for new Rrs data based on 
  default cluster settings or user-defined clusters (trained by FCM.new).
}
\examples{
library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)

}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_to_image}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
