% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{apply_FCM_m}
\alias{apply_FCM_m}
\title{Apply FCM_m to new input Rrs}
\usage{
apply_FCM_m(Rrs, wavelength, Rrs_clusters,
  stand=FALSE, default.cluster=TRUE, m_used=1.36, option.plot=FALSE)
}
\arguments{
\item{Rrs}{Data.frame, the input Rrs of FCM.}

\item{wavelength}{Numeric vector, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{Rrs_clusters}{Data.frame, used for applying FCM.
Default use the data from \code{Bi_clusters.rda}}

\item{stand}{Logical, whether to normalized the Rrs data. Default as \code{FALSE} means do not.}

\item{default.cluster}{Logical, whether to use the default clusters.
Default use the data from \code{Bi_clusters.rda}}

\item{m_used}{Number, Used fuzzifier value}

\item{option.plot}{Logical, whether to plot the result. Default as \code{FALSE}}
}
\value{
A \code{list} including several results of function \code{apply_FCM_m()}
  \itemize{
    \item \strong{x}  The raw input Rrs dataframe with unit sr^-1
    \item \strong{x.stand}  The standardized Rrs dataframe, if \code{stand=F}
    \item \strong{d}  Distance to each cluster
    \item \strong{u}  Membership values
    \item \strong{Area} Spectral intergration of each sample
    \item \strong{cluster}  Defined by the maximum of membership
    \item \strong{quality}  The quality of the cluster results.
    \item \strong{m.used}  The used value of fuzzifier(m)
    \item \strong{K}  Cluster number
    \item \strong{p.group}  A ggplot list for plotting the cluster result
    \item \strong{p.group.facet} \code{p.group} with facet to see each cluster resutls
      more clearly
    \item \strong{dt.plot}  Dataframe used for ggplot
  }
}
\description{
Application of FCM_m method for new Rrs data based on default cluster settings
  or user-defined clusters (trained by FCM.new).
}
\examples{
\dontrun{
library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)
}

}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_to_image}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
