% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_application.R
\name{apply_to_image}
\alias{apply_to_image}
\title{Apply the improved Fuzzy Cluster Method (FCMm) to raster data}
\usage{
apply_to_image(
  input,
  res,
  color_palette = RdYlBu(res$K),
  pos_rgb = NULL,
  output_image = FALSE,
  output_resultpng = FALSE,
  output_imRrs.n = FALSE,
  output_dir = ".",
  output_format = "GTiff",
  Chla_est = FALSE,
  title.name = NULL,
  png_scale = 50,
  fn_memb = "output_membership",
  fn_cluster = "output_cluster",
  fn_imRrs.n = "output_imRrs_normalized",
  fn_rgbpng = "output_rgb",
  fn_Chla = "output_Chla"
)
}
\arguments{
\item{input}{A \strong{raster} or a \strong{character} linking
to the raster file on the disk.}

\item{res}{A required list that used for clustering the image data including:
\itemize{
  \item \strong{FD}  The results from function \code{FuzzifierDetermination}
    providing wavelength
  \item \strong{K}  Cluster number
  \item \strong{res.FCM}  Cluster center and used fuzzifier value.
}
For the convenience, function \link{generate_param} supports to quickly
  generate this \code{list}. See more in examples.}

\item{color_palette}{The palette of cluster color. Default as \code{RdYlBu(res$K)}.
In \code{FCMm}, it could be \link{RdYlBu}, \link{Spectral}, \link{HUE} or other color values
with same length of cluster number.}

\item{pos_rgb}{The position of RGB channels, default as \code{NULL} to search the nearest band with 
665 (n), 560 (g), and 443 (b). Also could be a vector with three elements meaning the positions of 
required bands.}

\item{output_image}{Logical, whether to save images as raster files to your disk. Default as \code{FALSE}.}

\item{output_resultpng}{Logical, whether to save png files to your disk. Default as \code{FALSE}.}

\item{output_imRrs.n}{Logical, whether to produce normalized Rrs files. Default as \code{FALSE}. 
This parameter is used for inspection of FCM training. Will be \strong{deprecated} in the following version.}

\item{output_dir}{The directory of output files. Default as the current working directory (\code{"."})}

\item{output_format}{A string, the format of raster file, default as \code{"GTiff"}.
See more in \code{\link[raster:writeRaster]{writeRaster}}}

\item{Chla_est}{Logical, whether to estimate Chla concentration using the function \link{FCM_m_Chla_estimation}.
Default as \code{FALSE}.}

\item{title.name}{Character, the title name of ggplot for plotting. Default as \code{NULL}.}

\item{png_scale}{Numeric, scale of png. Recommended as \code{50}.}

\item{fn_memb}{A string, filename of membership raster file. Default as \code{"output_membership"}.}

\item{fn_cluster}{A string, filename of cluster raster file. Default as \code{"output_cluster"}.}

\item{fn_imRrs.n}{A string, filename of normalized Rrs raster file. Default as \code{"output_imRrs_normalized"}.}

\item{fn_rgbpng}{A string, file name of rgb png. Default as \code{"output_rgb"}.}

\item{fn_Chla}{A string, file name of estimated Chla raster file. Default as \code{"output_Chla"}.}
}
\value{
A \code{list()} of all results and several inputs:
  \itemize{
    \item \strong{input}  What we input character link to image file or raster object
    \item \strong{res}  Condition input of apply_to_image
    
    \item \strong{res.FCM}  List of FCM.new result
    \item \strong{res.Chla}  Data.frame includes the template and final Chla estimation results
    
    \item \strong{coord}  Dataframe of xy coordinates
    \item \strong{imRrs.raw}  Dataframe of raw Rrs excluding NA pixels. Thus, it may have less pixels than 
      that of \code{input}
    \item \strong{imRrs.n}  Dataframe of normalized Rrs. Same as \code{imRrs.raw} but in the normalized scale
    \item \strong{im.cluster} Dataframe of cluster with coordinates.
    \item \strong{im.memb} Dataframe of membership returned by \link{apply_FCM_m}. It includes coordinates.
    
    \item \strong{raster.memb}  Raster object of membership value. \code{NULL} if \code{output_image} is \code{FALSE}.
    \item \strong{raster.cluster}  Raster object of cluster. \code{NULL} if \code{output_image} is \code{FALSE}.
    \item \strong{raster.Chla} Raster object of Chla concentration. \code{NULL} if \code{output_image} is \code{FALSE}.
    
    \item \strong{p.memb}  A ggplot list of membership value map
    \item \strong{p.cluster}  A ggplot list of cluster map
    \item \strong{p.rgb} A ggplot list of rgb map
    \item \strong{p.Chla}  A ggplot list of Chla map
  }
}
\description{
This function could apply the defined water cluster to corrected image files.
  Should run \link{generate_param} or \link{generate_param_ex} to generate a \code{res} list as an input
  of function \code{apply_to_image}
}
\note{
\strong{2019-12-12}: 

  The section FCM running used the subset of default Rrs clusters.
  Please see the section \strong{Part II: New coming raster data} by running \code{vignette('Builtin_centrodis')}
  if have not known how to do in that situation.
  
  Also, if it is your first time to get the image data into \strong{R}, you could
  load the raster data by typing \code{raster::brick(filename)} which filename
  is the path of your data such as '/data/Test.dat' or 'E://data//Test.tiff' or so.
  
  \strong{2020-06-27}: 
  
  Note that the inputs of \code{apply_to_image} support the file path of raster file.
  
  \strong{2020-07-03}:
  
  Reported by Xiaolan Cai, the required RGB bands for plotting the true color image are defined by 
    inputted wavelength now.
}
\examples{
\donttest{
library(FCMm)
data("OLCI_TH")
data("Bi_clusters")
res <- generate_param(c(413,443,490,510,560,620,665,674,709,754,865,885))
im_result <- apply_to_image(input=OLCI_TH, res=res, title.name="Test_image", Chla_est=TRUE)
}
  
}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{cal_memb}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
