% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Assessment_via_cluster}
\alias{Assessment_via_cluster}
\title{Assessment each algorithm for every cluster}
\usage{
Assessment_via_cluster(
  pred,
  meas,
  memb,
  metrics = c("MAE", "MAPE"),
  log10 = FALSE,
  total = TRUE,
  hard.mode = TRUE,
  cal.precision = FALSE,
  na.process = FALSE,
  plot.col = FALSE
)
}
\arguments{
\item{pred}{prediciton of Chla}

\item{meas}{in-situ measurement of Chla}

\item{memb}{membership value matrix}

\item{metrics}{metrics need to be calculated}

\item{log10}{Should pred and meas be log10-transformed? (Default as FALSE)}

\item{total}{logical}

\item{hard.mode}{If \code{FALSE}, the membership values are used to calculate validation metrics}

\item{cal.precision}{Whether to calculate precision (only support for vectorized metrics), default as \code{FALSE}}

\item{na.process}{na.process and choose to statistic na value percent}

\item{plot.col}{option to plot col result for chosed metrics (Default as FALSE)}
}
\value{
List
}
\description{
Assessment each algorithm for every cluster
}
\note{
If the cal.precision is running, the \code{hard.mode == TRUE} is used. In that case,
  mean and sd calculation is conducted for hard mode based on result from cal.metrics.vector
}
\seealso{
Other Algorithm assessment: 
\code{\link{Getting_Asses_result}},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_interval}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
