% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{Assessment_via_cluster}
\alias{Assessment_via_cluster}
\title{Assessment each algorithm for every cluster}
\usage{
Assessment_via_cluster(
  pred,
  meas,
  memb,
  metrics = c("MAE2", "MAPE"),
  total = TRUE,
  hard.mode = TRUE
)
}
\arguments{
\item{pred}{prediciton of Chla}

\item{meas}{in-situ measurement of Chla}

\item{memb}{membership value matrix}

\item{metrics}{metrics need to be calculated}

\item{total}{logical}

\item{hard.mode}{hard.mode}
}
\value{
List
}
\description{
Assessment each algorithm for every cluster
}
