% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cal.metrics.names}
\alias{cal.metrics.names}
\title{List all metrics in function cal.metrics}
\usage{
cal.metrics.names()
}
\value{
strings
}
\description{
List all metrics in function cal.metrics
}
\seealso{
Other Utils: 
\code{\link{.level_to_variable}()},
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{lmodel2_}()}
}
\concept{Utils}
