% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cal.metrics}
\alias{cal.metrics}
\title{cal.metrics}
\usage{
cal.metrics(x,y,name="all")
}
\arguments{
\item{x}{True value}

\item{y}{Estimated value}

\item{name}{The name of metrics}
}
\description{
cal.metrics
}
