% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cal.metrics}
\alias{cal.metrics}
\title{Assess the performance of algorithms}
\usage{
cal.metrics(x,y,name="all",log10=FALSE)
}
\arguments{
\item{x}{True value OR Actual value}

\item{y}{Estimated value OR Predict value}

\item{name}{The name of metrics}

\item{log10}{Logical. Whether the input x and y should be log10-transformed}
}
\value{
Result of \code{cal.metrics} is a list of selected metric values.
}
\description{
Assess the performance of algorithms
}
\note{
(2020-02-09) All functions used log10 transformation was assigned to the 
  Key parameter `log10`.
}
\examples{
\dontrun{
set.seed(1234)
x = runif(20)
y = runif(20)
result = cal.metrics(x, y)
print(result)
}
}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{level_to_variable}()},
\code{\link{trim_sd}()}
}
\concept{Utils}
