% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.level_to_variable}
\alias{.level_to_variable}
\title{.level_to_variable}
\usage{
.level_to_variable(dt, warn = F)
}
\arguments{
\item{dt}{dataframe}

\item{warn}{warnning option}
}
\description{
.level_to_variable
}
