% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.level_to_variable}
\alias{.level_to_variable}
\title{Convert dataframe with factor to character}
\usage{
.level_to_variable(dt, warn = F)
}
\arguments{
\item{dt}{dataframe}

\item{warn}{warnning option}
}
\description{
Convert dataframe with factor to character
}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{lmodel2_}()}
}
\concept{Utils}
