% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{FBA_Le13}
\alias{FBA_Le13}
\title{FBA_Le13}
\usage{
FBA_Le13(Rrs665, Rrs682, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs682}{Rrs682}

\item{Rrs709}{Rrs709}
}
\description{
FBA_Le13
}
