% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation 2.R, R/Chla_estimation.R
\name{FBA_Yang10}
\alias{FBA_Yang10}
\title{Four-bands algorithm by Yang et al. (2010)}
\usage{
FBA_Yang10(Rrs665, Rrs709, Rrs754)

FBA_Yang10(Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\description{
Four-bands algorithm by Yang et al. (2010)

Four-bands algorithm by Yang et al. (2010)
}
\references{
Yang W, Matsushita B, Chen J, et al. An enhanced three-band index for estimating 
  chlorophyll-a in turbid case-II waters: case studies of Lake Kasumigaura, Japan, and Lake 
  Dianchi, China[J]. IEEE Geoscience and Remote Sensing Letters, 2010, 7(4): 655-659.

Yang W, Matsushita B, Chen J, et al. An enhanced three-band index for estimating 
  chlorophyll-a in turbid case-II waters: case studies of Lake Kasumigaura, Japan, and Lake 
  Dianchi, China[J]. IEEE Geoscience and Remote Sensing Letters, 2010, 7(4): 655-659.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}

Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
