% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{FCM.new}
\alias{FCM.new}
\title{Running the improved Fuzzy Cluster Method (FCM) by optimizing fuzzifier parameter}
\usage{
FCM.new(
  FDlist,
  K,
  sort.pos = length(FDlist$wv),
  sort.decreasing = FALSE,
  plot.jitter = TRUE,
  fast.mode = FALSE,
  stand = FALSE
)
}
\arguments{
\item{FDlist}{A \code{list} from function \code{\link{FuzzifierDetermination}}}

\item{K}{Number, cluster number}

\item{sort.pos}{The position to sort the cluster number. Default as the end position of the 
input training matrix. It can be \code{NULL} if there is no need to re-sort the cluster index.}

\item{sort.decreasing}{Default as \code{FALSE} which means the sort are assigned from small values
to large values.}

\item{plot.jitter}{Logical, choose to plot jitter plot using \code{ggplot2} functions}

\item{fast.mode}{Logical, \code{FALSE} (default)}

\item{stand}{Logical, choose to do normalization of input data}
}
\value{
A \code{list} of FCM:
  \itemize{
    \item \strong{FD} The return list by function \code{\link{FuzzifierDetermination}}
    \item \strong{res.FCM} The optimized FCM result generated by functions in package \code{ppclust}
    \item \strong{K} Cluster number
    \item \strong{centroids} The list of centroids (both at raw and normalized scale) of the cluster 
      result by aggregating each mean.
    \item \strong{plot.jitter} A logical value for the option of doing jitter plot by package \code{ggplot2}
    \item \strong{fast.mode} A logical value for choosing whether to use fast mode
  }
}
\description{
An improved version of Fuzzy Cluster Method (FCM) for water spectra data sets.
}
\examples{
library(FCMm) 
library(ggplot2) 
library(magrittr)
library(stringr)
data("Nechad2015")
x <- Nechad2015[,3:11]
wv <- gsub("X","",names(x)) \%>\% as.numeric
set.seed(1234)
w <- sample(1:nrow(x), 100) 
x <- x[w, ]
names(x) <- wv
nb = 4 # Obtained from the vignette "Cluster a new dataset by FCMm"
set.seed(1234)
FD <- FuzzifierDetermination(x, wv, stand=FALSE)
result <- FCM.new(FD, nb, fast.mode = TRUE)

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Dembele D. Multi-objective optimization for clustering 3-way gene
    expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3):
    211-225.
}
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{generate_param}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
