% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{FCM_m_Chla_estimation}
\alias{FCM_m_Chla_estimation}
\title{Estimate Chla concentration by algorithms blending via membership values}
\usage{
FCM_m_Chla_estimation(Rrs, U)
}
\arguments{
\item{Rrs}{Data.frame of Remote sensing reflectance which should
contain colname with (also at) \code{Rrs665}, \code{Rrs709},
and \code{Rrs754} with unit sr^-1.}

\item{U}{Data.frame of membership values which should have seven
columns presenting seven membership values from each cluster produced by FCM-m}
}
\value{
Each column presents Chla concentration with unit ug/L or mg/m^3 by model
  BR, TBA, C6 and blending result. Also with the membership values of each cluster.
}
\description{
To calculate the Chla concentration via blending remote sensing algorithms.
}
\note{
The input of \code{Rrs} must have bands with wavelength at 665, 709 and 754 nm.
  See examples of using this function.
}
\examples{
\dontrun{
library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)
dt_Chla <- FCM_m_Chla_estimation(Rrs=data.frame(Rrs665=Rrs$`665`, 
  Rrs709=Rrs$`708.75`,Rrs754=Rrs$`753.75`),U=result$u)
}

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Gilerson A A, Gitelson A A, Zhou J, et al. Algorithms for remote estimation 
    of chlorophyll-a in coastal and inland waters using red and near infrared bands[J].
    Optics Express, 2010, 18(23): 24109-24125.
}
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
