% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzifierDetermination.R
\name{FuzzifierDetermination}
\alias{FuzzifierDetermination}
\title{Determine the optimized fuzzifier for FCM running}
\usage{
FuzzifierDetermination(x, wv, max.m=10, stand=FALSE, dmetric='sqeuclidean')
}
\arguments{
\item{x}{Data.frame. the input Rrs data}

\item{wv}{Wavelength of X}

\item{max.m}{Set max.m as for determination of m.mub. Default as 10}

\item{stand}{Whether the data set is standardized (T needs not standardization)}

\item{dmetric}{Distance method. Default as 'sqeuclidean'}
}
\value{
\code{FD} list contains several result by \code{FuzzifierDetermination}:
  \itemize{
    \item \strong{x} The raw input Rrs dataframe with unit sr^-1
    \item \strong{x.stand} The standardized Rrs dataframe, if \code{stand=F}
    \item \strong{wv} Wavelength with unit nm
    \item \strong{max.m} The maximum fuzzifier of FCM as a restriction
    \item \strong{stand} A logic value for choosing whether to use standardization
    \item \strong{dmetric} A string value for choosing which distance metric to be used
    \item \strong{Area} A numeric vector for trapezoidal integral values
    \item \strong{m.ub} The upper boundary of fuzzifier(m) value
    \item \strong{m.used} The desired value of fuzzifier(m) value
  }
}
\description{
To determine the optimized fuzzifier value for FCM running.
}
\examples{
\dontrun{
library(FCMm)
library(tidyverse)
data("Nechad2015")
w <- Nechad2015 \%>\% names \%>\%
    str_extract(.,pattern="\\\\d") \%>\%
    is.na \%>\% {!.}
wv <- w \%>\% names(Nechad2015)[.] \%>\%
    gsub('X','',.) \%>\% as.numeric
x <- w \%>\% Nechad2015[,.]
names(x) <- wv
rm(w)
FD <- FuzzifierDetermination(x, wv, stand=F)
}

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
  \item Dembele D. Multi-objective optimization for clustering 3-way gene
    expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3):
    211-225.
}
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
