% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_application.R
\name{generate_param}
\alias{generate_param}
\alias{generate_param_ex}
\title{Generate the input parameter of apply_to_image based on 
  the built-in or user-defined centroids}
\usage{
generate_param(wl)

generate_param_ex(centroids, m = 1.36)
}
\arguments{
\item{wl}{wavelength of subsetting}

\item{centroids}{The centorids values required for fuzzy clustering on images. Its colnames should
be converted to numeric wavelength. The format should be like \code{Rrs_clusters.default}.}

\item{m}{The fuzzifier value of FCM. Default as \code{1.36}.}
}
\value{
The return of \code{generate_param} or \code{generate_param_ex} is a list used for 
  function \link{apply_to_image}.
}
\description{
Generate the input parameter of apply_to_image based on 
  the built-in or user-defined centroids
}
\note{
\code{generate_param} only support the cluster centroids proposed by \code{Bi et al. (2019)}.
  However, the \code{generate_param_ex} could be used for user-defined centroids if you want.
}
\examples{
library(FCMm)
wl = c(413, 443, 490, 510, 560, 620, 665, 674, 681, 709, 754, 779, 865, 885)
res = generate_param(wl)

}
\references{
\itemize{
  \item Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
    an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
}
\seealso{
Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{cal_memb}()},
\code{\link{plot_spec_from_df}()}
}
\concept{Fuzzy cluster functions}
