% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Getting_Asses_results}
\alias{Getting_Asses_results}
\title{Get the result of function Assessment_via_cluster}
\usage{
Getting_Asses_results(
  sample.size,
  replace = FALSE,
  pred,
  meas,
  memb,
  cluster,
  seed = NULL
)
}
\arguments{
\item{sample.size}{Sample size. This supports a bootstrap way to run the function 
\code{Assessment_via_cluster}. The number should not be larger than the row number 
of pred or so.}

\item{replace}{Logical, replace, default as \code{FALSE}}

\item{pred}{Prediction matrix}

\item{meas}{Measured (actual) vector}

\item{memb}{Membership matrix}

\item{cluster}{Cluster vector}

\item{seed}{Seed number for fixing the random process. See \code{help(set.seed)} for more details.}
}
\value{
A list containing fuzzy and hard results from \code{Assessment_via_cluster}
}
\description{
This function mainly use function \code{Assessment_via_cluster} to get
  assesments both from fuzzy and hard mode. Specifically, it will return fuzzy_MAE, 
  fuzzy_MAPE, hard_Slope, Valid_percent, and hard_Rsquare which should be seemed as the 
  input value of function \code{Scoring_system}.
}
\note{
The row number of \code{pred}, \code{meas}, \code{memb}, and \code{cluster} should be the same. 
  This function is designed for bootstrapping process to get Chla algorithms assessment. Therefore, 
  parameters of \code{Assessment_via_cluster} is set as fixed such as \code{log10 = TRUE}, 
  \code{na.process = TRUE}. Given that, I will not export this function in latter to avoid confuses.
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_interval}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
