% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{Gons08}
\alias{Gons08}
\title{Gons08}
\usage{
Gons08(Rrs665, Rrs709, Rrs779)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs779}{Rrs779}
}
\description{
Gons08
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{CI_Hu12}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
