% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{Gons08}
\alias{Gons08}
\title{Gons algorithm by Gons et al. (2008)}
\usage{
Gons08(Rrs665, Rrs709, Rrs779)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs779}{Rrs779}
}
\value{
A list including the algorithm estimated bbp and Chla concentration.
}
\description{
Gons algorithm by Gons et al. (2008)
}
\examples{
data(WaterSpec35)
res = Gons08(WaterSpec35$`665`, WaterSpec35$`708.75`, WaterSpec35$`778.75`)
}
\references{
Gons H J, Auer M T, Effler S W. MERIS satellite chlorophyll mapping of 
  oligotrophic and eutrophic waters in the Laurentian Great Lakes[J]. Remote Sensing 
  of Environment, 2008, 112(11): 4098-4106.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Bi21}()},
\code{\link{Blend_FCMm}()},
\code{\link{Blend_Jac17}()},
\code{\link{Blend_Moo14}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
