% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{NDCI_Mi12}
\alias{NDCI_Mi12}
\title{NDCI_Mi12}
\usage{
NDCI_Mi12(Rrs665, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}
}
\description{
NDCI_Mi12
}
\references{
Mishra S, Mishra D R. Normalized difference chlorophyll index:
  A novel model for remote estimation of chlorophyll-a concentration in turbid 
  productive waters[J]. Remote Sensing of Environment, 2012, 117: 394-406.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
