% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{NDCI_Mi12}
\alias{NDCI_Mi12}
\title{Normalized Difference Chlorophyll Index (NDCI) algorithm by Mishra et al. (2012)}
\usage{
NDCI_Mi12(Rrs665, Rrs709)
}
\arguments{
\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
Normalized Difference Chlorophyll Index (NDCI) algorithm by Mishra et al. (2012)
}
\note{
Chla_test as a test version of NDCI_Mi12 results parameterized by Bi
}
\examples{
data(WaterSpec35)
res = NDCI_Mi12(WaterSpec35$`665`, WaterSpec35$`708.75`)
}
\references{
Mishra S, Mishra D R. Normalized difference chlorophyll index:
  A novel model for remote estimation of chlorophyll-a concentration in turbid 
  productive waters[J]. Remote Sensing of Environment, 2012, 117: 394-406.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{Smith18}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
