% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{OCI_Hu12}
\alias{OCI_Hu12}
\title{Ocean Color Index (OCI) algorithm by Hu et al. (2012)}
\usage{
OCI_Hu12(Rrs443, Rrs490, Rrs510, Rrs560, Rrs665)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs490}{Rrs490}

\item{Rrs510}{Rrs510}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}
}
\value{
A list including the algorithm indices and Chla concentration.
}
\description{
Ocean Color Index (OCI) algorithm by Hu et al. (2012)
}
\examples{
data(WaterSpec35)
res = OCI_Hu12(WaterSpec35$`442.5`, WaterSpec35$`490`, WaterSpec35$`510`, 
WaterSpec35$`560`, WaterSpec35$`665`)
}
\references{
Hu C, Lee Z, Franz B. Chlorophyll a algorithms for oligotrophic oceans: 
  A novel approach based on three band reflectance difference[J]. Journal of Geophysical 
  Research: Oceans, 2012, 117(C1).
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Jac17}()},
\code{\link{Blend_Moo14}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
