% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{plot_spec}
\alias{plot_spec}
\title{Plot results of the improved Fuzzy Cluster Method (FCMm)}
\usage{
plot_spec(
  res,
  show.stand = NULL,
  show.ribbon = FALSE,
  color_palette = RdYlBu(res$K),
  prefix = "Cluster "
)
}
\arguments{
\item{res}{The result of function \link{FCM.new}}

\item{show.stand}{Whether to show the spectra on standarized scale (default as \code{NULL} 
and assigned by \code{res$FD$stand}).}

\item{show.ribbon}{Whether to show the ribbon defined by sd values calculated from samples 
in each cluster (default as \code{FALSE}).}

\item{color_palette}{The palette of cluster color. Default as \code{RdYlBu(res$K)}.
In \code{FCMm}, it could be \link{RdYlBu}, \link{Spectral}, \link{HUE} or other color values
with same length of cluster number.}
}
\value{
The result of \code{plot_spec} is a list including:
  \itemize{
    \item \strong{p.all.spec} Spectra of all data.
    \item \strong{p.cluster.spec} Spectra of clusters.
    \item \strong{p.group.spec.1} Spectra of groups in mode 1
    \item \strong{p.group.spec.2} Spectra of groups in mode 2
    \item \strong{res} Result list of \link{FCM.new}
    \item \strong{show.stand} Logical value whether use standardized y-axis
  }
}
\description{
Spectra plots for a \code{FCM.new} result
}
\examples{
library(FCMm) 
library(ggplot2) 
library(magrittr)
data("Nechad2015")
x <- Nechad2015[,3:11]
wv <- gsub("X","",names(x)) \%>\% as.numeric
set.seed(1234) # Set this seed so that you can re-produce them
w <- sample(1:nrow(x), 100)
x <- x[w, ]
names(x) <- wv
nb = 4 # Obtained from the vignette "Cluster a new dataset by FCMm"
set.seed(1234)
FD <- FuzzifierDetermination(x, wv, stand=FALSE)
result <- FCM.new(FD, nb, fast.mode = TRUE)
p.spec <- plot_spec(result, show.stand = TRUE)
print(p.spec$p.cluster.spec)

}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
