% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm 2.R, R/FCMm.R
\name{plot_spec}
\alias{plot_spec}
\title{Plot the result of FCM_m}
\usage{
plot_spec(res, show.stand=FALSE, HABc=NULL)

plot_spec(res, show.stand=FALSE, HABc=NULL)
}
\arguments{
\item{res}{The result of function `FCM.new`}

\item{show.stand}{Logical option to show the standardized spectra (default as False)}

\item{HABc}{Numeric value to specify which cluster owning larger magnitude}
}
\value{
A list of ggplot lists:
  \itemize{
    \item \strong{p.all.spec} Spectra of all data.
    \item \strong{p.cluster.spec} Spectra of clusters.
    \item \strong{p.group.spec.1} Spectra of groups in mode 1
    \item \strong{p.group.spec.2} Spectra of groups in mode 2
    \item \strong{res} Result list of \code{FCM.new}
    \item \strong{show.stand} Logical value whether use standardized y-axis
  }

A list of ggplot lists:
  \itemize{
    \item \strong{p.all.spec} Spectra of all data.
    \item \strong{p.cluster.spec} Spectra of clusters.
    \item \strong{p.group.spec.1} Spectra of groups in mode 1
    \item \strong{p.group.spec.2} Spectra of groups in mode 2
    \item \strong{res} Result list of \code{FCM.new}
    \item \strong{show.stand} Logical value whether use standardized y-axis
  }
}
\description{
Spectra plot for a FCM.new result

Spectra plot for a FCM.new result
}
\examples{
\dontrun{p.spec <- plot_spec(result, show.stand=FALSE, HABc=NULL)}

\dontrun{p.spec <- plot_spec(result, show.stand=FALSE, HABc=NULL)}

}
\references{
Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
  an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
}
