% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{plot_spec_from_df}
\alias{plot_spec_from_df}
\title{Plot spectra from dataframe}
\usage{
plot_spec_from_df(df)
}
\arguments{
\item{df}{Data.frame for Rrs plotting}
}
\value{
The result of \code{plot_spec_from_df} is a \code{ggplot} object showing spectra curve.
}
\description{
Spectra plot for a dataframe given numeric colnames
}
\note{
The colnames of input `df` should be numericalize.
}
\examples{
data(WaterSpec35)
plot_spec_from_df(WaterSpec35[, -c(1, 2)])
# which is equal to 
plot_spec_from_df(WaterSpec35)

}
\seealso{
plot_spec_group

Other Fuzzy cluster functions: 
\code{\link{FCM.new}()},
\code{\link{FuzzifierDetermination}()},
\code{\link{apply_FCM_m}()},
\code{\link{apply_to_image}()},
\code{\link{generate_param}()}
}
\concept{Fuzzy cluster functions}
