% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{QAA_v5}
\alias{QAA_v5}
\title{Quasi-Analytical Algorithm version 5 (QAA_v5) by Lee et al. (2002)}
\usage{
QAA_v5(
  wv,
  Rrs,
  wv412 = NULL,
  wv443 = NULL,
  wv490 = NULL,
  wv560 = NULL,
  wv667 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{wv}{Wavelength vector}

\item{Rrs}{Rrs matrix}

\item{wv412}{Wavelength index of 412 nm}

\item{wv443}{Wavelength index of 443 nm}

\item{wv490}{Wavelength index of 490 nm}

\item{wv560}{Wavelength index of 560 nm}

\item{wv667}{Wavelength index of 667 nm}

\item{verbose}{Print process information (default as FASLE)}
}
\value{
Results of \code{QAA_v5} are the list including:
  \item{input}{input list including \code{wv}, \code{Rrs}, \code{wv412},
    \code{wv443}, \code{wv490}, \code{wv560}, and \code{wv667}}
  \item{const}{constant values of \code{g0}, \code{g1}, and \code{lam0}}
  \item{Var}{Indirect variables including \code{rrs}, \code{u}, \code{Chi},
    \code{Yita}, \code{Zeta}, \code{S}, and \code{Xi}}
  \item{IOP}{Inherent optical properties including \code{a0}, \code{bbp0},
    \code{ag443}, \code{a}, \code{adg}, \code{aph}, and \code{bbp}}
  \item{Chl}{Results of Chl concentration estimation including \code{Chl_Bricaud}, 
    \code{Bricaud_A}, \code{Bricaud_B}, \code{Chl_Liu}, \code{aChla665_star}}
  \item{residu}{Residual constrains including \code{Upper_lim_for_Rrs667}, \code{Lower_lim_for_Rrs667},
    \code{Rrs667_revised}, and \code{Rrs667_final}}
}
\description{
Quasi-Analytical Algorithm version 5 (QAA_v5) by Lee et al. (2002)
}
\examples{
data(WaterSpec35)
Rrs = WaterSpec35[,-c(1,2)]
wv = as.numeric(names(Rrs))
print(wv)
res = QAA_v5(wv, Rrs, wv412 = 2, wv443 = 3, wv490 = 4, wv560 = 6, wv667 = 9)

}
\references{
Lee Z P, Carder K L, Arnone R A. Deriving inherent optical properties from water color: 
  a multiband quasi-analytical algorithm for optically deep waters[J]. Applied optics, 2002, 41(27): 
  5755-5772.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Blend_Bi21}()},
\code{\link{Blend_FCMm}()},
\code{\link{Blend_Jac17}()},
\code{\link{Blend_Moo14}()},
\code{\link{Blend_Smith18}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OCx_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{SCI_Shen10}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
