% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_srf_excel}
\alias{read_srf_excel}
\title{Read srf data from excel file}
\usage{
read_srf_excel(fn)
}
\arguments{
\item{fn}{filename of excel file}
}
\value{
The list of SRF for different sensors.
}
\description{
Read srf data from excel file
}
\examples{
\dontrun{
# Suppose fn is a MS spread file of which sheetnames as sensor names.
# For each sheet, it includes the data with colnames as bands, 
#   rownames as wavelength, and cell values as SRF.
SRF_LIST = read_srf_excel(fn)
}
}
