% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rrs_clusters.default.R
\docType{data}
\name{Rrs_clusters.default}
\alias{Rrs_clusters.default}
\title{Rrs_clusters.default}
\format{dataframe with 7 rows by 15 cols}
\usage{
Rrs_clusters.default
}
\description{
A dataframe contains the water optical clusters produced by Bi et al. (2019)
}
\keyword{datasets}
