% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{run_all_Chla_algorithms}
\alias{run_all_Chla_algorithms}
\title{Run all Chla algorithms except for `QAA_v5`}
\usage{
run_all_Chla_algorithms(Rrs, wv_range = 3)
}
\arguments{
\item{Rrs}{Dataframe that should with required colnames 443, 490, 510, 560, 620, 665, 681, 709, 754, 779}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}
}
\value{
A list including Chla results by supported algorithms if Rrs includes the required wavelength.
  Retrun \code{NA} values for algorithms of which required wavelength cannot be found.
}
\description{
Run all Chla algorithms except for `QAA_v5`
}
\details{
Please type \code{Chla_algorithm_name()} to see all Chla algorithms
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()}
}
\concept{Algorithms: Chla concentration}
