% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{run_all_Chla_algorithms}
\alias{run_all_Chla_algorithms}
\title{run_all_Chla_algorithms}
\usage{
run_all_Chla_algorithms(Rrs, wv_range = 3)
}
\arguments{
\item{Rrs}{Dataframe that should with required colnames 443, 490, 510, 560, 620, 665, 681, 709, 754, 779}

\item{wv_range}{Number that used to define the range of wavelength to capture
the center wavelength of required band}
}
\value{
A list
}
\description{
run_all_Chla_algorithms
}
\details{
Please type \code{Chal_algorithm_name()} to see all Chla algorithms
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4E}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()}
}
\concept{Algorithms: Chla concentration}
