% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{SCI_Shen10}
\alias{SCI_Shen10}
\title{SCI_Shen10}
\usage{
SCI_Shen10(Rrs560, Rrs620, Rrs665, Rrs681)
}
\arguments{
\item{Rrs560}{Rrs560}

\item{Rrs620}{Rrs620}

\item{Rrs665}{Rrs665}

\item{Rrs681}{Rrs681}
}
\description{
SCI_Shen10
}
\references{
Shen F, Zhou Y X, Li D J, et al. Medium resolution imaging spectrometer 
  (MERIS) estimation of chlorophyll-a concentration in the turbid sediment-laden waters 
  of the Changjiang (Yangtze) Estuary[J]. International Journal of Remote Sensing, 2010, 
  31(17-18): 4635-4650.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
