% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{SCI_Shen10}
\alias{SCI_Shen10}
\title{Synthetic Chlorophyll Index (SCI) algorithm by Shen et al. (2010)}
\usage{
SCI_Shen10(Rrs560, Rrs620, Rrs665, Rrs681)
}
\arguments{
\item{Rrs560}{Rrs560}

\item{Rrs620}{Rrs620}

\item{Rrs665}{Rrs665}

\item{Rrs681}{Rrs681}
}
\value{
A list including the algorithm index and Chla concentration.
}
\description{
Synthetic Chlorophyll Index (SCI) algorithm by Shen et al. (2010)
}
\examples{
data(WaterSpec35)
res = SCI_Shen10(WaterSpec35$`560`, WaterSpec35$`620`, 
  WaterSpec35$`665`, WaterSpec35$`681.25`)
}
\references{
Shen F, Zhou Y X, Li D J, et al. Medium resolution imaging spectrometer 
  (MERIS) estimation of chlorophyll-a concentration in the turbid sediment-laden waters 
  of the Changjiang (Yangtze) Estuary[J]. International Journal of Remote Sensing, 2010, 
  31(17-18): 4635-4650.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{Smith18}()},
\code{\link{TC2}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
