% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Score_algorithms_interval}
\alias{Score_algorithms_interval}
\title{Score a vector of error metrics for algorithms}
\usage{
Score_algorithms_interval(
  x,
  trim = FALSE,
  reward.punishment = TRUE,
  decreasing = TRUE,
  hundred.percent = FALSE
)
}
\arguments{
\item{x}{Input vector of error metrics (NA values are allowed)}

\item{trim}{whether to run a trim process to calculate mean and standard deviation of 
input vector x (Default as \code{FALSE})}

\item{reward.punishment}{wheter to conduct the reward and punishment mechainism in scoring
system (Default as \code{TRUE})}

\item{decreasing}{the order of the good metric to be evaluated. For instance, MAE should use
\code{decreasing = TRUE} (Default) since the algorithm performs better when MAE becomes smaller. 
However, when comes to \code{Rsquare} from linear regression (maximum is 1), it should be
\code{FALSE}}

\item{hundred.percent}{A variable constrain for metrics that the maximun of input \code{x} 
should not be greater than 100.}
}
\value{
List
}
\description{
Score a vector of error metrics for algorithms
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
