% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Score_algorithms_sort}
\alias{Score_algorithms_sort}
\title{Score a vector of error metrics for algorithms by the sort}
\usage{
Score_algorithms_sort(x, decreasing = TRUE)
}
\arguments{
\item{x}{Input vector of error metrics (NA values are allowed)}

\item{decreasing}{the order of the good metric to be evaluated. For instance, MAE should use
\code{decreasing = TRUE} (Default) since the algorithm performs better when MAE becomes smaller. 
However, when comes to \code{Rsquare} from linear regression (maximum is 1), it should be
\code{FALSE}}
}
\value{
Results of \code{Score_algorithms_sort()} are returned as a vector presenting score values.
}
\description{
Score a vector of error metrics for algorithms by the sort
}
\examples{
\dontrun{
set.seed(1234)
x = runif(10)
result = Score_algorithms_sort(x)
}
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_results}()},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms_interval}()},
\code{\link{Scoring_system}()}
}
\concept{Algorithm assessment}
