% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms_assessment.R
\name{Scoring_system}
\alias{Scoring_system}
\title{The main function for algorithms scoring}
\usage{
Scoring_system(Inputs, trim = FALSE, remove.negative = FALSE)
}
\arguments{
\item{Inputs}{The list returned form function \code{Getting_Asses_results}}

\item{trim}{The input parameter of function \code{Score_algorithms} (Default as \code{FALSE})}

\item{remove.negative}{Option to repalce the negative score as zero (Default as \code{FALSE})}
}
\value{
A list including Total_score, Accuracy, Precision, Effectiveness, Accuracy.list, 
  Precision.list, and Total_score.melt.
}
\description{
The main function for algorithms scoring
}
\seealso{
Other Algorithm assessment: 
\code{\link{Assessment_via_cluster}()},
\code{\link{Getting_Asses_result}},
\code{\link{Sampling_via_cluster}()},
\code{\link{Score_algorithms}()}
}
\concept{Algorithm assessment}
