% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{SRF_simulate}
\alias{SRF_simulate}
\title{Simulation of hyperspectral data by SRF}
\usage{
SRF_simulate(
  Rrs,
  select_sensor = "All",
  output_wavelength = "MED",
  save_as_csv = FALSE,
  na.rm = TRUE,
  wv_as_column = TRUE
)
}
\arguments{
\item{Rrs}{A data.frame with colnames as Wavelength + SampleName, the first column is wavelength.}

\item{select_sensor}{Character. Select sensors. Use \code{show_sensor_names()} to print
all supported sensors. Default as \code{All}}

\item{output_wavelength}{Character. \code{MED} (default) or \code{MAX}.
Define the center wavelength. \code{MED} means the center wavelength is 
middle position of half maximun of max peak. While \code{MAX} means the 
position at the maximun of SRF peak.}

\item{save_as_csv}{Logical. Choose to save the simulation results as single csv for each
sensor. Default with \code{FALSE}}

\item{na.rm}{Logical. Should NA values be removed? Default as \code{TRUE}}

\item{wv_as_column}{Logical. If \code{TRUE} (default), the output result is a dataframe
with wavelength as column names.}
}
\value{
A \code{list} with names as all selected sensors from parameters \code{select_sensor}.
For each list, including five elements:
  \itemize{
    \item \strong{sensor} Sensor name
    \item \strong{srf} Spectral response function of the sensor
    \item \strong{cw_med} Center wavelength by method \code{MED}
    \item \strong{cw_max} Center wavelength by method \code{MAX}
    \item \strong{Rrs_simu} The simulation of Rrs by supported SRF
  }
}
\description{
Simulate hyperspectral Rrs to multispectral bands
  via sensors SRF (Spectral response function).
}
\examples{
\dontrun{
library(FCMm)
nm <- seq(400,900)
Rrs <- data.frame(nm=nm,Site1=exp(nm/1000)+runif(501))
result <- SRF_simulate(Rrs,select_sensor=c("OLCI","MODIS"))
}

}
\seealso{
Other Utils: 
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{level_to_variable}()},
\code{\link{lmodel2_}()},
\code{\link{trim_sd}()}
}
\concept{Utils}
