% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chla_estimation.R
\name{TC2}
\alias{TC2}
\title{Chla concentration algorithm for Turbid Case-2 (TC2) waters by Liu et al. (2020)}
\usage{
TC2(Rrs443, Rrs560, Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
A list including MCI index, Chla_final, Chla_clean, Chla_turbid, and flag.
}
\description{
Chla concentration algorithm for Turbid Case-2 (TC2) waters by Liu et al. (2020)
}
\examples{
data(WaterSpec35)
res = TC2(WaterSpec35$`442.5`, WaterSpec35$`560`, WaterSpec35$`665`,
WaterSpec35$`708.75`, WaterSpec35$`753.75`)

}
\references{
Liu G, Li L, Song K, et al. An OLCI-based algorithm for semi-empirically 
  partitioning absorption coefficient and estimating chlorophyll a concentration in 
  various turbid case-2 waters[J]. Remote Sensing of Environment, 2020, 239: 111648.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{Bloom}()},
\code{\link{C6}()},
\code{\link{Chla_FBA}},
\code{\link{Chla_Gil10_Git11}},
\code{\link{Chla_OC_OLCI}},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
