% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM_m_Chla_estimation.R
\name{TC2}
\alias{TC2}
\title{TC2}
\usage{
TC2(Rrs443, Rrs560, Rrs665, Rrs709, Rrs754)
}
\arguments{
\item{Rrs443}{Rrs443}

\item{Rrs560}{Rrs560}

\item{Rrs665}{Rrs665}

\item{Rrs709}{Rrs709}

\item{Rrs754}{Rrs754}
}
\value{
A list with names as Chla_final, Chla_clean, Chla_turbid, and flag
}
\description{
TC2
}
\references{
Liu G, Li L, Song K, et al. An OLCI-based algorithm for semi-empirically 
  partitioning absorption coefficient and estimating chlorophyll a concentration in 
  various turbid case-2 waters[J]. Remote Sensing of Environment, 2020, 239: 111648.
}
\seealso{
Other Algorithms: Chla concentration: 
\code{\link{BR_Gil10}()},
\code{\link{BR_Git11}()},
\code{\link{C6}()},
\code{\link{Chla_algorithms_name}()},
\code{\link{FBA_Le13}()},
\code{\link{FBA_Yang10}()},
\code{\link{FCM_m_Chla_estimation}()},
\code{\link{Gons08}()},
\code{\link{NDCI_Mi12}()},
\code{\link{OC4_OLCI}()},
\code{\link{OC5_OLCI}()},
\code{\link{OC6_OLCI}()},
\code{\link{OCI_Hu12}()},
\code{\link{QAA_v5}()},
\code{\link{SCI_Shen10}()},
\code{\link{TBA_Gil10}()},
\code{\link{TBA_Git11}()},
\code{\link{TC2_clean}()},
\code{\link{TC2_turbid}()},
\code{\link{run_all_Chla_algorithms}()}
}
\concept{Algorithms: Chla concentration}
