% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMm.R
\name{trapz}
\alias{trapz}
\title{Trapezoid integral calculation}
\usage{
trapz(wv, x)
}
\arguments{
\item{wv}{wavelength}

\item{x}{data.frame that need to be trapzed with ncol equal to length of\code{wv}}
}
\value{
A vector presents the area result.
}
\description{
Trapezoid integral calculation
}
\examples{
\dontrun{
library(FCMm)
library(tidyverse)
data("Nechad2015")
w <- Nechad2015 \%>\% names \%>\%
    str_extract(.,pattern="\\\\d") \%>\%
    is.na \%>\% {!.}
wv <- w \%>\% names(Nechad2015)[.] \%>\%
    gsub('X','',.) \%>\% as.numeric
x <- w \%>\% Nechad2015[,.]
names(x) <- wv
rm(w)
Area <- trapz(wv, x)
}
}
