% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{trim_sd}
\alias{trim_sd}
\title{Standard deviation by trimmed values}
\usage{
trim_sd(x, trim = 0.05, na.rm = T)
}
\arguments{
\item{x}{input numeric vector}

\item{trim}{percent of length(x) to be trimmed}

\item{na.rm}{default as TRUE to remove NA values}
}
\value{
sd value
}
\description{
Standard deviation by trimmed values
}
\seealso{
Other Utils: 
\code{\link{SRF_simulate}()},
\code{\link{cal.metrics.names}()},
\code{\link{cal.metrics.vector.names}()},
\code{\link{cal.metrics.vector}()},
\code{\link{cal.metrics}()},
\code{\link{level_to_variable}()},
\code{\link{lmodel2_}()}
}
\concept{Utils}
